requirements.txt
Flask==3.0.0
Pillow==10.0.0
numpy==1.26.0
matplotlib==3.8.0
gunicorn==21.2.0